idrec/requirements.txt
numpy
pillow
pandas
streamlit
openai
